# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Foodie API
  version: 1.0.0
  description: |
    Peer-to-peer homemade-food marketplace API. Customers browse dishes from
    home kitchens, place pickup orders, and leave reviews; food makers manage
    their kitchen, menu, orders, and earnings; admins oversee the platform.
    Money is in integer cents. Auth is a Bearer JWT from /api/auth/*.
servers:
  - url: http://localhost:4000
    description: Local dev
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderStatus:
      type: string
      enum: [NEW, ACTIVE, COMPLETED, CANCELLED]
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: { type: array, items: { type: object } }
...
    OrderItem:
      type: object
      properties:
        id: { type: string }
        dishId: { type: string }
        nameSnapshot: { type: string }
        priceCentsSnapshot: { type: integer }
        qty: { type: integer }
    Order:
      type: object
      properties:
        id: { type: string, example: FD123456 }
        customerId: { type: string }
        kitchenId: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        subtotalCents: { type: integer }
        serviceFeeCents: { type: integer }
        totalCents: { type: integer }
        paymentMethod: { type: string }
        pickupSlot: { type: string }
        paymentIntentId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        items: { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
    Review:
      type: object
      properties:
        id: { type: string }
        orderId: { type: string }
        customerId: { type: string }
        kitchenId: { type: string }
        overall: { type: integer, minimum: 1, maximum: 5 }
        quality: { type: integer, minimum: 1, maximum: 5 }
        packaging: { type: integer, minimum: 1, maximum: 5 }
        accuracy: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
...
    Paginated:
      type: object
      required: [data, page, limit, total]
      properties:
        data: { type: array, items: { type: object } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
paths:
...
  /api/orders:
    post:
      tags: [orders]
      summary: Place an order (all items must be from one kitchen)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, paymentMethod, pickupSlot]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [dishId, qty]
                    properties:
                      dishId: { type: string, example: chicken-biryani }
                      qty: { type: integer, minimum: 1, maximum: 20, example: 2 }
                paymentMethod: { type: string, example: 'Apple Pay' }
                pickupSlot: { type: string, example: '5:00 PM - 7:00 PM' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: '#/components/schemas/Order' }
                  paymentIntent: { type: object, description: 'Mocked until Stripe is wired' }
        '400': { description: Mixed kitchens, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Dish not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Insufficient stock / unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/orders/mine:
    get:
      tags: [orders]
      summary: My orders, newest first
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Order' } } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Order detail (customer, kitchen owner, or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/orders/{id}/review:
    post:
      tags: [orders]
      summary: Review a completed order (one per order)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [overall, quality, packaging, accuracy]
              properties:
                overall: { type: integer, minimum: 1, maximum: 5 }
                quality: { type: integer, minimum: 1, maximum: 5 }
                packaging: { type: integer, minimum: 1, maximum: 5 }
                accuracy: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Review' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Not your order, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Not completed / already reviewed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/maker/orders:
    get:
      tags: [maker]
      summary: Incoming orders for your kitchen
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Order' } } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Requires MAKER role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/orders/{id}:
    patch:
      tags: [maker]
      summary: Accept (NEW→ACTIVE) or reject (NEW→CANCELLED, stock restored) an order
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [accept, reject] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Order is not NEW, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/admin/orders:
    get:
      tags: [admin]
      summary: Paginated order list with optional status filter (ADMIN role)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } } }
        '403': { description: Requires ADMIN role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...